requirements.txt
flask
gunicorn
scikit-learn
tensorflow-cpu
numpy
pandas